Finished the required logic, following the description.
I was hitting a few walls when I went back to follow the assignment step
by step. I had the fundamental logic down, but I was getting stuck when
trying to write the function which would play the game 5 times. I couldn't
get my head round how I could use the playerchoice parameter multiple
times, as it was a prompt, and I had declared at the top of the page.

I tried using the prompt in the function as a parameter but this didn't
work. Turned out it was mainly an issue of order. I had written the prompt
at the top of the document after the computerchoice function. Seems obvious
now, but because the parameter of the function had the same name as the
variable, I thought I had to assign the value to the variable before I
called the function.

On to the next!